+++ alkymi/config.py
@@ -1,4 +1,7 @@
 from enum import Enum
+from pathlib import Path
+import os
+from typing import Optional
 
 
 class CacheType(Enum):
@@ -35,6 +38,7 @@ class AlkymiConfig:
         # Set default values in config
         AlkymiConfig.__instance = self
         self._cache = True  # type: bool
+        self._cache_path = None  # type: Optional[Path]
 
     @property
     def cache(self) -> bool:
@@ -51,6 +55,26 @@ class AlkymiConfig:
         :param enable_cache: Whether to enable alkymi caching globally
         """
         self._cache = enable_cache
+
+    @property
+    def cache_path(self) -> Optional[Path]:
+        """
+        :return: A user-provided location to place the cache
+        """
+        return self._cache_path
+
+    @cache_path.setter
+    def cache_path(self, cache_path: Path) -> None:
+        """
+        Set a custom cache path to override the default caching location
+
+        :param cache_path: The custom location to place the cache
+        """
+        if not cache_path.is_dir():
+            raise ValueError("Cache path '{}' must be a directory".format(cache_path))
+        if not os.access(str(cache_path), os.W_OK):
+            raise ValueError("Cache path '{}' must be writeable".format(cache_path))
+        self._cache_path = cache_path
 
 
 # Force creation of singleton