+++ CHANGELOG.md
@@ -9,17 +9,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated README.md with "Command Line Usage", "Upcoming Features" and "Known Issues" sections
 - Included `labfile.py` in static type checking and linting
 - Added a unit test for the built-in `file()` recipe generator
+- Made cache directory configurable using `AlkymiConfig` and added a test for it
+- Added a new test to verify that partially executed ForeachRecipes can be reloaded from cache and continue from point
+of interruption/failure
 
 ### Removed
 - "clean" and "clean-cache" commands from Lab (these will be reimplemented properly at a later point)
 
 ### Changed
 - Renamed alkymi's labfile from `lab.py` to `labfile.py` to fit the new "labfile" naming scheme
+- ForeachRecipe now supports saving/restoring partial evaluation states to the cache. This fixes the issue where a
+failure in one of the foreach evaluations could cause all the work to be lost
+- Recipes will now be marked dirty if their bound function has changed between invocations
 
 ### Fixed
 - Fixed type annotations in alkymi's labfile (`labfile.py`)
 - Fixed reloading of recipe checksums from cache (these needed to be converted from lists to tuples due to JSON
 serialization format). This would cause some recipes to be marked dirty even though they weren't.
+- Fixed a bug where the aggregate checksum for all mapped inputs to a ForeachRecipe weren't being saved and restored
+from cache
 
 ## [0.0.3] - 2021-01-03
 ### Added